Add mapOptions prop to Collection component
Feature to provide optional settings for the behaviour of the map that displays the spatial extent of the collection.

I named it `mapOptions` and not `leafletOptions` on purpose so that it's library-independent. That's also the reason for naming `noWrap` differently (it's not a very speaking property name anyway and I don't like the negation).