Improve documentation of SupportedFileFormats
The term "output file formats" shouldn't appear in the general description of the component anymore.
Also, I think there should be a note that the expected structure of the `formats` prop differs depending on the value of `version`. I tried to express this with "as defined by the respective version" -- I'm not 100% happy with that wording yet, but I think it's better than before.